HSEARCH-4303 Refactor automatic indexing association tests
So that we can add tests for more association types
