+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/bytype/SingleValuedPropertyAccessor.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype;
 
 import java.util.function.BiConsumer;
 import java.util.function.Function;
