+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/association/bytype/manytomany/ownedbycontaining/AutomaticIndexingManyToManyOwnedByContainingListBaseIT.java
@@ -4,12 +4,10 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.multi;
+package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.manytomany.ownedbycontaining;
 
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Stream;
 import javax.persistence.Basic;
@@ -21,20 +19,15 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToMany;
-import javax.persistence.MapKeyClass;
-import javax.persistence.MapKeyJoinColumn;
 import javax.persistence.OneToOne;
 import javax.persistence.OrderBy;
 import javax.persistence.Transient;
 
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.AbstractAutomaticIndexingAssociationIT;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.ContainerPrimitives;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.MultiValuedPropertyAccessor;
-import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.SingleValuedPropertyAccessor;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.AbstractAutomaticIndexingMultiValuedAssociationBaseIT;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.ContainerPrimitives;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.MultiValuedPropertyAccessor;
+import org.hibernate.search.integrationtest.mapper.orm.automaticindexing.association.bytype.SingleValuedPropertyAccessor;
 import org.hibernate.search.mapper.pojo.automaticindexing.ReindexOnUpdate;
-import org.hibernate.search.mapper.pojo.extractor.builtin.BuiltinContainerExtractors;
-import org.hibernate.search.mapper.pojo.mapping.definition.annotation.AssociationInverseSide;
-import org.hibernate.search.mapper.pojo.extractor.mapping.annotation.ContainerExtraction;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.IndexedEmbedded;
@@ -43,29 +36,26 @@ import org.hibernate.search.mapper.pojo.mapping.definition.annotation.ObjectPath
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.PropertyValue;
 
 /**
- * Test automatic indexing based on Hibernate ORM entity events
- * when a Map-keys association is involved.
- * <p>
- * See {@link AbstractAutomaticIndexingAssociationIT} for more details on how this test is designed.
+ * Test automatic indexing caused by multi-valued association updates
+ * or by updates of associated (contained) entities,
+ * with a {@code @ManyToMany List} association owned by the containing side.
  */
-public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticIndexingMultiAssociationIT<
-		AutomaticIndexingMapKeysAssociationIT.IndexedEntity,
-		AutomaticIndexingMapKeysAssociationIT.ContainingEntity,
-		AutomaticIndexingMapKeysAssociationIT.ContainedEntity,
-		Map<AutomaticIndexingMapKeysAssociationIT.ContainedEntity, String>,
-		List<AutomaticIndexingMapKeysAssociationIT.ContainingEntity>
-		> {
+public class AutomaticIndexingManyToManyOwnedByContainingListBaseIT
+		extends AbstractAutomaticIndexingMultiValuedAssociationBaseIT<
+						AutomaticIndexingManyToManyOwnedByContainingListBaseIT.IndexedEntity,
+						AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainingEntity,
+						AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainedEntity,
+						List<AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainedEntity>,
+						List<AutomaticIndexingManyToManyOwnedByContainingListBaseIT.ContainingEntity>
+				> {
 
-	public AutomaticIndexingMapKeysAssociationIT() {
-		super( new ModelPrimitives() );
+	public AutomaticIndexingManyToManyOwnedByContainingListBaseIT() {
+		super( new ModelPrimitivesImpl() );
 	}
 
-	private static class ModelPrimitives
-			implements MultiAssociationModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity,
-						Map<ContainedEntity, String>, List<ContainingEntity>> {
-
-		private final ContainerPrimitives<Map<ContainedEntity, String>, ContainedEntity> MAP_KEYS_PRIMITIVES =
-				ContainerPrimitives.mapKeys( containedEntity -> containedEntity.getIndexedField() + "_value" );
+	private static class ModelPrimitivesImpl
+			implements MultiValuedModelPrimitives<IndexedEntity, ContainingEntity, ContainedEntity,
+					List<ContainedEntity>, List<ContainingEntity>> {
 
 		@Override
 		public String getIndexName() {
@@ -125,13 +115,13 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public Map<ContainedEntity, String> newContainedAssociation(Map<ContainedEntity, String> original) {
-			return new LinkedHashMap<>( original );
+		public List<ContainedEntity> newContainedAssociation(List<ContainedEntity> original) {
+			return new ArrayList<>( original );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbedded() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbedded() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbedded, ContainingEntity::setContainedIndexedEmbedded );
 		}
 
@@ -142,8 +132,8 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedNonIndexedEmbedded() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedNonIndexedEmbedded() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedNonIndexedEmbedded, ContainingEntity::setContainedNonIndexedEmbedded );
 		}
 
@@ -154,23 +144,24 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbeddedShallowReindexOnUpdate() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbeddedShallowReindexOnUpdate() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbeddedShallowReindexOnUpdate,
 					ContainingEntity::setContainedIndexedEmbeddedShallowReindexOnUpdate );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbeddedNoReindexOnUpdate() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbeddedNoReindexOnUpdate() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
 					ContainingEntity::getContainedIndexedEmbeddedNoReindexOnUpdate,
 					ContainingEntity::setContainedIndexedEmbeddedNoReindexOnUpdate );
 		}
 
 		@Override
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedUsedInCrossEntityDerivedProperty() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
-					ContainingEntity::getContainedUsedInCrossEntityDerivedProperty );
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedUsedInCrossEntityDerivedProperty() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
+					ContainingEntity::getContainedUsedInCrossEntityDerivedProperty,
+					ContainingEntity::setContainedUsedInCrossEntityDerivedProperty );
 		}
 
 		@Override
@@ -181,9 +172,9 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 
 		@Override
 		@SuppressWarnings("unchecked")
-		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, Map<ContainedEntity, String>> containedIndexedEmbeddedWithCast() {
-			return new MultiValuedPropertyAccessor<>( MAP_KEYS_PRIMITIVES,
-					root -> (Map) root.getContainedIndexedEmbeddedWithCast() );
+		public MultiValuedPropertyAccessor<ContainingEntity, ContainedEntity, List<ContainedEntity>> containedIndexedEmbeddedWithCast() {
+			return new MultiValuedPropertyAccessor<>( ContainerPrimitives.collection(),
+					root -> (List) root.getContainedIndexedEmbeddedWithCast() );
 		}
 
 		@Override
@@ -265,91 +256,40 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		})
 		private ContainingEntity child;
 
-		@ElementCollection
-		@JoinTable(
-				name = "i_containedIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbedded = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "i_containedIndexedEmbedded")
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		private List<ContainedEntity> containedIndexedEmbedded = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "i_containedNonIndexedEmbedded",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<ContainedEntity, String> containedNonIndexedEmbedded = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "i_containedNonIndexedEmbedded", joinColumns = @JoinColumn(name = "containingNonIndexedEmbedded"),
+			inverseJoinColumns = @JoinColumn(name = "containedNonIndexedEmbedded"))
+		private List<ContainedEntity> containedNonIndexedEmbedded = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "i_indexedEmbeddedShallow",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		@IndexingDependency(
-				reindexOnUpdate = ReindexOnUpdate.SHALLOW,
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbeddedShallowReindexOnUpdate = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "i_indexedEmbeddedShallow",
+			inverseJoinColumns = @JoinColumn(name = "indexedEmbeddedShallow"))
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.SHALLOW)
+		private List<ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "i_indexedEmbeddedNoReindex",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		@IndexingDependency(
-				reindexOnUpdate = ReindexOnUpdate.NO,
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-		)
-		private Map<ContainedEntity, String> containedIndexedEmbeddedNoReindexOnUpdate = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "i_indexedEmbeddedNoReindex",
+			inverseJoinColumns = @JoinColumn(name = "indexedEmbeddedNoReindex"))
+		@IndexedEmbedded(includePaths = { "indexedField", "indexedElementCollectionField", "containedDerivedField" })
+		@IndexingDependency(reindexOnUpdate = ReindexOnUpdate.NO)
+		private List<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "i_containedCrossEntityDP",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		private Map<ContainedEntity, String> containedUsedInCrossEntityDerivedProperty = new LinkedHashMap<>();
+		@ManyToMany
+		@JoinTable(name = "i_containedCrossEntityDP", joinColumns = @JoinColumn(name = "containingCrossEntityDP"),
+				inverseJoinColumns = @JoinColumn(name = "containedCrossEntityDP"))
+		private List<ContainedEntity> containedUsedInCrossEntityDerivedProperty = new ArrayList<>();
 
-		@ElementCollection
-		@JoinTable(
-				name = "i_containedIndexedEmbeddedCast",
-				joinColumns = @JoinColumn(name = "mapHolder")
-		)
-		@MapKeyClass(ContainedEntity.class)
-		@MapKeyJoinColumn(name = "map_key")
-		@Column(name = "value")
-		@OrderBy("map_key asc") // Forces Hibernate ORM to use a LinkedHashMap; we make sure to insert entries in the correct order
-		@IndexedEmbedded(
-				includePaths = { "indexedField" },
-				extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY),
-				targetType = ContainedEntity.class
-		)
-		private Map<Object, String> containedIndexedEmbeddedWithCast = new LinkedHashMap<>();
+		@ManyToMany(targetEntity = ContainedEntity.class)
+		@JoinTable(name = "i_containedIndexedEmbeddedCast", joinColumns = @JoinColumn(name = "containingIndexedEmbeddedCast"),
+				inverseJoinColumns = @JoinColumn(name = "containedIndexedEmbeddedCast"))
+		@IndexedEmbedded(includePaths = "indexedField", targetType = ContainedEntity.class)
+		private List<Object> containedIndexedEmbeddedWithCast = new ArrayList<>();
 
 		public Integer getId() {
 			return id;
@@ -383,45 +323,50 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 			this.child = child;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbedded() {
+		public List<ContainedEntity> getContainedIndexedEmbedded() {
 			return containedIndexedEmbedded;
 		}
 
-		public void setContainedIndexedEmbedded(Map<ContainedEntity, String> containedIndexedEmbedded) {
+		public void setContainedIndexedEmbedded(List<ContainedEntity> containedIndexedEmbedded) {
 			this.containedIndexedEmbedded = containedIndexedEmbedded;
 		}
 
-		public Map<ContainedEntity, String> getContainedNonIndexedEmbedded() {
+		public List<ContainedEntity> getContainedNonIndexedEmbedded() {
 			return containedNonIndexedEmbedded;
 		}
 
-		public void setContainedNonIndexedEmbedded(Map<ContainedEntity, String> containedNonIndexedEmbedded) {
+		public void setContainedNonIndexedEmbedded(List<ContainedEntity> containedNonIndexedEmbedded) {
 			this.containedNonIndexedEmbedded = containedNonIndexedEmbedded;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbeddedShallowReindexOnUpdate() {
+		public List<ContainedEntity> getContainedIndexedEmbeddedShallowReindexOnUpdate() {
 			return containedIndexedEmbeddedShallowReindexOnUpdate;
 		}
 
 		public void setContainedIndexedEmbeddedShallowReindexOnUpdate(
-				Map<ContainedEntity, String> containedIndexedEmbeddedShallowReindexOnUpdate) {
+				List<ContainedEntity> containedIndexedEmbeddedShallowReindexOnUpdate) {
 			this.containedIndexedEmbeddedShallowReindexOnUpdate = containedIndexedEmbeddedShallowReindexOnUpdate;
 		}
 
-		public Map<ContainedEntity, String> getContainedIndexedEmbeddedNoReindexOnUpdate() {
+		public List<ContainedEntity> getContainedIndexedEmbeddedNoReindexOnUpdate() {
 			return containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
 		public void setContainedIndexedEmbeddedNoReindexOnUpdate(
-				Map<ContainedEntity, String> containedIndexedEmbeddedNoReindexOnUpdate) {
+				List<ContainedEntity> containedIndexedEmbeddedNoReindexOnUpdate) {
 			this.containedIndexedEmbeddedNoReindexOnUpdate = containedIndexedEmbeddedNoReindexOnUpdate;
 		}
 
-		public Map<ContainedEntity, String> getContainedUsedInCrossEntityDerivedProperty() {
+		public List<ContainedEntity> getContainedUsedInCrossEntityDerivedProperty() {
 			return containedUsedInCrossEntityDerivedProperty;
 		}
 
-		public Map<Object, String> getContainedIndexedEmbeddedWithCast() {
+		public void setContainedUsedInCrossEntityDerivedProperty(
+				List<ContainedEntity> containedUsedInCrossEntityDerivedProperty) {
+			this.containedUsedInCrossEntityDerivedProperty = containedUsedInCrossEntityDerivedProperty;
+		}
+
+		public List<Object> getContainedIndexedEmbeddedWithCast() {
 			return containedIndexedEmbeddedWithCast;
 		}
 
@@ -429,23 +374,17 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		@GenericField
 		@IndexingDependency(derivedFrom = {
 				@ObjectPath({
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						),
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
 						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField1")
 				}),
 				@ObjectPath({
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						),
+						@PropertyValue(propertyName = "containedUsedInCrossEntityDerivedProperty"),
 						@PropertyValue(propertyName = "fieldUsedInCrossEntityDerivedField2")
 				})
 		})
 		public Optional<String> getCrossEntityDerivedField() {
-			return AbstractAutomaticIndexingAssociationIT.computeDerived(
-					containedUsedInCrossEntityDerivedProperty.keySet().stream().flatMap( c -> Stream.of(
+			return computeDerived(
+					containedUsedInCrossEntityDerivedProperty.stream().flatMap( c -> Stream.of(
 							c.getFieldUsedInCrossEntityDerivedField1(),
 							c.getFieldUsedInCrossEntityDerivedField2()
 					) )
@@ -467,66 +406,20 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 		@Id
 		private Integer id;
 
-		/*
-		 * No mappedBy here. The inverse side of associations modeled by a Map key cannot use mappedBy.
-		 * If they do, Hibernate assumes that map *values* are the opposite side of the association,
-		 * and ends up adding all kind of wrong foreign keys.
-		 */
-		@ManyToMany
-		@JoinTable(name = "contained_mapHolder")
+		@ManyToMany(mappedBy = "containedIndexedEmbedded")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedIndexedEmbedded",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<ContainingEntity> containingAsIndexedEmbedded = new ArrayList<>();
 
-		/*
-		 * No mappedBy here, same reasons as above.
-		 */
-		@ManyToMany
-		@JoinTable(name = "contained_nonIndexedMapHolder",
-				inverseJoinColumns = @JoinColumn(name = "containingNonIndexedEmbedded"))
+		@ManyToMany(mappedBy = "containedNonIndexedEmbedded")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
 		private List<ContainingEntity> containingAsNonIndexedEmbedded = new ArrayList<>();
 
-		/*
-		 * No mappedBy here, same reasons as above.
-		 */
-		@ManyToMany
-		@JoinTable(inverseJoinColumns = @JoinColumn(name = "containingCrossEntityPD"))
+		@ManyToMany(mappedBy = "containedUsedInCrossEntityDerivedProperty")
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedUsedInCrossEntityDerivedProperty",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<ContainingEntity> containingAsUsedInCrossEntityDerivedProperty = new ArrayList<>();
 
-		/*
-		 * No mappedBy here. The inverse side of associations modeled by a Map key cannot use mappedBy.
-		 * If they do, Hibernate assumes that map *values* are the opposite side of the association,
-		 * and ends up adding all kind of wrong foreign keys.
-		 */
-		@ManyToMany(targetEntity = ContainingEntity.class)
-		@JoinTable(name = "contained_withCastMapHolder",
-			inverseJoinColumns = @JoinColumn(name = "containingIndexedEmbeddedCast"))
+		@ManyToMany(mappedBy = "containedIndexedEmbeddedWithCast", targetEntity = ContainingEntity.class)
 		@OrderBy("id asc") // Make sure the iteration order is predictable
-		@AssociationInverseSide(
-				inversePath = @ObjectPath(
-						@PropertyValue(
-								propertyName = "containedIndexedEmbeddedWithCast",
-								extraction = @ContainerExtraction(BuiltinContainerExtractors.MAP_KEY)
-						)
-				)
-		)
 		private List<Object> containingAsIndexedEmbeddedWithCast = new ArrayList<>();
 
 		@Basic
@@ -661,8 +554,7 @@ public class AutomaticIndexingMapKeysAssociationIT extends AbstractAutomaticInde
 				@ObjectPath(@PropertyValue(propertyName = "fieldUsedInContainedDerivedField2"))
 		})
 		public Optional<String> getContainedDerivedField() {
-			return AbstractAutomaticIndexingAssociationIT.computeDerived( Stream.of( fieldUsedInContainedDerivedField1, fieldUsedInContainedDerivedField2 ) );
+			return computeDerived( Stream.of( fieldUsedInContainedDerivedField1, fieldUsedInContainedDerivedField2 ) );
 		}
 	}
-
 }
